added generation for missing values
